pyproject.toml
[project]
name = "new-site"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Maxime Filippini", email = "maxime.filippini@gmail.com" }
]
dependencies = [
    "jinja2>=3.1.3",
    "mistune>=3.0.2",
    "PyYAML>=6.0.1",
    "pydantic>=2.6.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
hello = "new_site:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "djLint>=1.34.1",
    "pytest>=8.0.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/new_site"]

[tool.pytest.ini_options]
testpaths = [
    "tests",
]